requirements.txt
psutil==6.1.0
ceo-py==0.5.5b0
python-dotenv==1.0.1
sympy==1.13.3